requirements.txt
Brotli==1.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
diffusers==0.33.1
filelock==3.18.0
fonttools==4.58.0
fsspec==2025.5.1
hf-xet==1.1.2
huggingface-hub==0.32.2
idna==3.10
imagecodecs==2025.3.30
importlib_metadata==8.7.0
inflate64==1.0.1
iniconfig==2.1.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
multivolumefile==0.2.3
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==25.0
pillow==11.2.1
pluggy==1.6.0
psutil==7.0.0
py7zr==0.22.0
pybcj==1.0.6
pycryptodomex==3.23.0
pyparsing==3.2.3
pyppmd==1.1.1
pytest==8.3.5
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyzstd==0.17.0
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
six==1.17.0
sympy==1.14.0
texttable==1.7.0
tifffile==2025.5.26
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
triton==3.3.0
typing_extensions==4.13.2
urllib3==2.4.0
zipp==3.22.0
